requirements.txt
flask
flask-login
flask-mail
flask-sqlalchemy
flask-migrate
flask-whooshalchemy
flask-wtf
flask-babel
coverage
qrcode[pil]
xhtml2pdf
matplotlib
